Fix leaking resources, Fix #776

+++ src/main/java/net/openhft/chronicle/queue/BenchmarkMain.java
@@ -68,7 +68,7 @@ public class BenchmarkMain {
             ExcerptTailer tailer = queue.createTailer().toEnd();
             long endLoop = System.nanoTime();
             while (running) {
-                loopTime.sample(System.nanoTime() - endLoop);
+                loopTime.sample((double) (System.nanoTime() - endLoop));
                 Jvm.safepoint();
 
 //                    readerLoopTime = System.nanoTime();
@@ -162,8 +162,8 @@ public class BenchmarkMain {
             Bytes<?> bytes = wire.bytes();
             long start = readMessage(bytes);
             long end = System.nanoTime();
-            transportTime.sample(transport - start);
-            readTime.sample(end - transport);
+            transportTime.sample((double) (transport - start));
+            readTime.sample((double) (end - transport));
         }
         Jvm.safepoint();
     }
